server/requirements.txt
openai-whisper
transformers
ffmpeg
torch
torchaudio
fastapi[all]
uvicorn
streamlink
